pyproject.toml
[project]
name = "p5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "joblib>=1.5.3",
    "jupyter>=1.1.1",
    "lab>=8.8",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
[tool.pytest.ini_options]
pythonpath = ["."]
[tool.coverage.run]
omit = [
    "database/create_db.py",
    "database/insert_data.py",
    "api/__init__.py",
    "database/__init__.py"
]